backend/requirements.txt
django
djangorestframework
django-filter
django-extensions
django-rest-registration
social-auth-app-django
markdown
psycopg2
flake8
ipython
ipdb
spotipy
pillow
django-cors-headers
celery
redis
requests